Revert "Merge pull request #6172 from Cacodemon345/armfixes-2"
This reverts commit 5bfa69f3da072b4e6fdf8d0e9e937e5a084b8963, reversing
changes made to a54e33bf041e3a0cb311d51e5d97e0d41320d246.

+++ src/codegen_new/codegen_ir_defs.h
@@ -336,7 +336,7 @@ typedef struct uop_t {
     ir_reg_t      src_reg_a;
     ir_reg_t      src_reg_b;
     ir_reg_t      src_reg_c;
-    uintptr_t     imm_data;
+    uint32_t      imm_data;
     void         *p;
     ir_host_reg_t dest_reg_a_real;
     ir_host_reg_t src_reg_a_real, src_reg_b_real, src_reg_c_real;
@@ -601,7 +601,7 @@ uop_gen_reg_src3_imm(uint32_t uop_type, ir_data_t *ir, int src_reg_a, int src_re
 }
 
 static inline void
-uop_gen_imm(uint32_t uop_type, ir_data_t *ir, uintptr_t imm)
+uop_gen_imm(uint32_t uop_type, ir_data_t *ir, uint32_t imm)
 {
     uop_t *uop = uop_alloc(ir, uop_type);
